pyproject.toml
[tool.poetry]
name = "device_certificate_report"
version = "0.1.0"
description = "A tool to generate device certificate reports from PAN-OS devices."
authors = ["Calvin Remsburg <cremsburg.dev@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "device_certificate_report"}]

[tool.poetry.dependencies]
python = "^3.12"
pan-os-python = "^1.12.1"
dynaconf = "^3.2.6"
typer = "^0.12.5"
pydantic = "^2.9.2"
setuptools = "^75.1.0"
reportlab = "^4.2.2"


[tool.poetry.group.dev.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.35"
pytest = "^8.3.3"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
device_certificate_report = "device_certificate_report.main:app"
